Controller: wait for device initialization
After creating all the soft interfaces specified in the recipe, the
Controller will now wait until the associated Device object on the Slave
is initialized - this means waiting until there are no more devices left
in the _tmp_mapping dictionary of InterfaceManager.

To do this we introduced a new NetTestSlave method - wait_interface_init
that only returns after the InterfaceManager is ready.

This method is also called every time you create a device from inside a
python task.

This should avoid potential race conditions that we've experienced.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Tested-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>