InterfaceAPI: Allow user to configure STP state on interface
When the kernel's STP is not enabled it's possible for the user to
explictly set the STP state of a netdev. This can be useful for
switchdev driver authors who would like to make sure their driver acts
accordingly.

The STP state is expressed as a number in order to be consistent with
the underlying 'bridge' utility:
	0 - port is DISABLED
	1 - STP LISTENING state
	2 - STP LEARNING state
	3 - STP FORWARDING state
	4 - STP BLOCKING state

Acked-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>