SlavePool: sort machines and interfaces in matching algorithm
We want to have matching algorithm deterministic. For that we need to
secure, that data are entered in the same order each time. This patch
adds sorting of machines by their id's before they enter matching algorithm
so it shouldn't happen that for same setup, machines are mapped differently
in more runs

Version 2: Sort also interfaces and use reverse sorting just to make sure
we're processing the hostnames and interfaces in alphabetical order
while matching.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Jan Tluka <jtluka@redhat.com>