lnst: Fix dynamic addition / deletion of slaves
The slaves of an interface are determined during creation time and
stored in two objects:
	* Interface on the controller
	* NetConfigDeviceGeneric on the slave machine

When slaves are dynamically added / deleted during a Python task the
'slaves' data structures in these objects are not updated, which can
cause several problems:

If a slave is removed from its master, then during cleanup the slave
machine will try to remove it again and thereby generate an error. This
requires us to update the data struct on the slave machine.

If a slave is added to a master and that master is later reset(), then
the newly added slave won't be configured correctly after the reset().
This requires us to update the data struct on the controller.

Cc: Andrew Lunn <andrew@lunn.ch>
Fixes: af025186a329 ("InterfaceAPI: Introduce slave_{add, del}")
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>