recipes: switchdev: Test LAG leave flow
When port netdevs are leaving a bridged LAG device, then the bridge's
cleanup sequence isn't invoked and the switch driver needs to do the
necessary cleanup.

Adjust current LAG recipes to test the flow.

Note that I'm creating a new LAG device on the switch for the second
fastpath test, as currently switchdev drivers suffer from an ordering
problem, which first requires the LAG to be populated and only then
bridged. Hopefully, this will be resolved soon.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>