Config: add security section and options to config
This adds some basic security parameters that will be used by the
lnst-slave and lnst-ctl applications.

On the Controller it's just 2 options:
identity - a name that the controller uses to identify to the Slave,
    Slave stores public keys of controllers with their identification
privkey - the path to the private key of the Controller

Both of options are only used in case we're using the pubkey
authetication of the Slave.

On the Slave it's these options:
auth_types - the accepted authentication types, at this moment it only
    takes a single value but in the future it will be a list of
    authentication types provided to Controllers. Possible values are:
    none, password, pubkey, ssh
auth_password - if auth_types == password then this password will be
    used for authentication
privkey - path to the Slaves private key that will be used when
    auth_types == pubkey
ctl_pubkeys - path to directory where the Slave should look for public
    keys of Controllers when auth_types == pubkey

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>