Add SecureSocket classes
Added 3 new classes: SecureSocket, CtlSecSocket and SlaveSecSocket. The
SecureSocket class encapsulates a normal socket object and adds what is
basically a TLS Record Layer - adds a signature, padding and encrypts
the entire message.

The CtlSecSocket and SlaveSecSocket classes are child classes inheriting
from the SecureSocket class and implement the handshake protocols used
from the Controller or from the Slave side. The important method is
handshake() that takes a sec_params parameter which is a dictionary
describing how to do the handshake.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>