Use SecureSockets for Ctl<->Slave communication
This patch replaces the sockets used for the Controller<->Slave
communication with CtlSecSocket and SlaveSecSocket. This includes some
changes to the ConnectionHandler by adding SecureSocket support and
removing the socket support since it's not needed anymore (at least for
now).

Caution: Right now this breaks the functionality of the lnst-ctl
deconfigure command and the lnst-pool-wizard, which are at the moment
not able to use the secured sockets.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>