{Ctl, Slave}SecSocket: add insecure secure channel type
This renames the old "none" auth_type to "no-auth" and adds a new "none"
auth_type that works the same way as we've had until now - no
authentication in place. This also adds some warning log messages
indicating the insecurity of the communication channel to both the
controller and the slave.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>