SecureSocket: make cryptography import optional
The cryptography library is imported only when authentication methods
that need it are requested. If the library is missing, the user will get
an error message explaining the problem. It's a pretty hackish way of
doing the imports but I'm not aware of a better solution... I'm open to
suggestions if there are any.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>