NetTestController: deconfigure now uses SecureSockets
This updates the save_machine_config method to store the security
information of the machine into the temporary file. This information is
later used to connect to the machine during the deconfiguration. This
can probably be done in a more sensible way, but that would probably
require reimplementation of these two methods to work in a different
way. To add at least some form of security I made sure that the
temporary file only has rw- permissions for the owner.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>