Wizard: add support for SecureSockets
This makes the Wizard use the CtlSecureSocket instead of the basic
socket. In case of noninteractive mode it is automatically assumed that
the communication between Slave and the wizard is not secured - it uses
the "none" authentication method.

In case of interactive mode, the user will be asked to provide security
parameters that will be used for the handshake. These will also be
included in the resulting slave machine description XML file.

If the connection fails during the handshake the user will be notified
of this.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>