RecipeCommon: added PerfRepo helper functions module for python tasks
This module provides functions that are related to PerfRepo. Currently
contains just one function generate_perfrepo_comment() that returns
string containing various information such as kernel versions on the
hosts passed as argument, if running in Beaker environment the job url
and additional user specified string to be included in comment. This
function will be used in phase1/phase2 regression tests.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>