TaskAPI: fix exception when perfrepo credentials are not in config
When user does not have PerfRepo installed and tries to use LNST PerfRepo
commands, user is informed, PerfRepo commands are ignored and test is run
without them.

When user does have PerfRepo installed, but does not have url/username/password
informations in config file, LNST crashes on PerfRepoRESTAPIException
because it tries to connect to URL "".

Desired behaviour would be - inform user, that he does not have
needed credentials in config and ignore the commands in the same
way they are ignored when the PerfRepo is not installed on the
machine.

This patch fixes that. User gets warning logging messages for all
of credentials that are missing and that PerfRepo commands will be
ignored and test is run without PerfRepo commands.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>