XmlParser: allow use HTTPS URL as recipe file
Current lxml.etree.parse implementation allows using HTTP, but not
HTTPS. When user tries to use HTTPS URL, it crashes on IOException.

This patch fixes that by checking if the URL starts with HTTPS
and uses urlopen to open the HTTPS page. This allows lxml.etree.parse
method to parse the XML from the HTTPS URL.

urllib2 is already used Common/Path and is available since Python2.6
so no new dependency is added.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>