TaskAPI: fix saving of the result after wait intr and kill
I believe this was a typo in the past. If a command is run in background
and later intr(), kill() or wait() is called on the command the result
was saved in never used self._res variable. The ProcessAPI object has
self._res_cmd variable that keeps the result of a command. Once the command
on background finishes (intr, wait, kill) it should be updated by the
collected data from associated intr, wait, kill commands.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>