SecureSocket: use fallback function if bit_length is not available in Python
Recently added SecureSocket feature brought dependency on Python 2.7
namely bit_length() method. Since we need to keep compatibility with
Python 2.6 that's on RHEL6 I added a fallback if bit_length is not
available. I compared the DH_GROUP and SRP_GROUP dicts both on Python2.6
and 2.7 and they had the same content (bit length), so this patch should
be safe.

Fixes issue 166.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>