NetTestController: fix syncing of tools with mixed python and xml tasks
When python task is in the recipe xml with other xml tasks specified
before, the  sync_table gets overwritten by default resources.

This is because the res_table is copied from the default resources
in the loop for every task in the recipe.

The fix is to move initialization of res_table before the loop.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>