Netperf: print human readable values of measured rate
Currently Netperf module transforms each rate into bps internally.
When reporting the rates/thresholds like this is not user friendly.

I changed the reporting so that rate is reported like this:

Measured rate 858.66 +-0.00 Mbits/sec is higher than threshold 700.00 +-0.00 Mbits/sec

instead of this:

Measured rate 914154070.00 +-0.00 bps is higher than threshold 600.00 +-100.00

The data in res_data is left intact, the change is only for
logging.

Signed-off-by: Jan Tluka <jtluka@redhat.com>