Netperf: fixes for rate pretty printing
If threshold is not specified the test will fail with an exception.
The fix is to process threshold values for prettier output only if
it was specified.

Also if threshold is not specified the rates were printed in bps.
Now they are printed using pretty rates.

This patch also fixes pretty print of the deviation.

Signed-off-by: Jan Tluka <jtluka@redhat.com>