InterfaceManager: handle RTM_{NEW, DEL}ADDR nl_messages
InterfaceManager and Device objects are now capable of handling ip
address related netlink messages. The Device object stores addresses in
a list as dictionaries. This should change in the future to either use
the Python ipaddress module or to use our own custom objects for ip
addresses.

Additionally the Device objects now generate an update message for the
controller in the init_netlink method as well and we'll be sending an
update message for all devices, even ones that aren't mapped based on
the currently running recipe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>