InterfaceManager: extend Device::if_data method
The if_data method now returns a complete dictionary describing the
Device object.

This includes a nonbackwards compatible change of renaming "devname" to
just "name".

v2: peer attribute is a reference to a Device object. The Controller
can't use it and it's not pickleable. So send just the if_index of the
peer

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>