NetTestSlave: remake get_devices and add get_device
The get_device slave method now returns a list of ALL devices that the
InterfaceManager sees and their data returned by the recently updated
get_if_data method.

This commit also adds a get_device method that returns the result of
get_if_data for one single Device, identified by the interface index in
the kernel.

Additionally this updates the Wizard class as well since it's the only
one that was using the get_devices method. It now uses the
get_devices_by_params method and selects just the DOWN && eth
interfaces.

This is a backwards incompatible change.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>