Machine: construct a Device database reflecting InterfaceManager state
This commit adds a new Device class to the Machine module. This class
reflects the Device class from the InterfaceManager wrt. values received
from the kernel and works as a synchronized getter/setter for them.

Right now the Device class only implements the get methods and just the
set_mtu method. This is because the whole idea of this class conflicts
with the Interface classes that we already have and is therefore very
hackish and should be reworked in the future.

This hackish solution is needed here while we still use XML Recipes and
want to test some more complex OvS features, such as VxLan tunneling.

After merging Python recipes this will be completely reworked and Device
objects (similar to how they're now) should completely replace the
Interface objects which are created based on XML requirements.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>