Task: add DeviceAPI as extension of Machine device database
The DeviceAPI exports the Device objects stored in the device database
of a machine to the python task. For now it only implements get methods
for Device information received from the Slave, and the set_mtu method.

The implemented methods are compatible with relevant InterfaceAPI
methods.

This patch also adds an optional 'selector' argument to the get_ip and
get_ips methods of InterfaceAPI, this argument is not used and is here
only so that these methods are compatible with DeviceAPI methods.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>