InterfaceManager: add Device objects to tmp_mapping sooner
This makes sure that the _is_name_used sees the Device object while it
is being created. This is so that if the Device creation actually
creates more than one interface at once it doesn't reuse the same name
for them. An example of such a case is creation of an OvS bridge with
internal interfaces.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>