InterfaceManager: rescan devices after reconnecting to netlink
When we reconnect to netlink we risk bringing the InterfaceManager into
an inconsistent state by discarding broadcast messages from the old
socket. Because of this we should call rescan_devices every time a new
netlink socket is bound.

Calling this method manually in the NetTestSlave after namespace
creation is therefore not necessary anymore.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>