InterfaceManager: bind to only interesting groups
The default bind() call on the IPRSocket binds us to all possible
RTNL broadcast groups. This restricts those groups to just those related
to link and ip address updates. This should reduce the number of
messages, hopefuly lowering the chances of the socket running out of
buffer space and getting closed by the kernel.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>