NetTestSlave: reconnect netlink from ServerHandler
The ServerHandler automatically closes all sockets that are either
closed from the other side or get into a broken state. This is normally
fine, however in case the socket in question is the Netlink socket, the
InterfaceManager doesn't get notified of this event and attempts to call
on this object fail with EBADF.

This makes sure that whenever the ServerHandler closes the socket, it
calls the InterfaceManager reconnect_netlink() method so that a new
netlink connection is established.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>