InterfaceManager, Machine: replace viewitems with items
The viewitems dictionary method was introduced in Python2.7. This
breaks our setups on RHEL6 where only Python2.6 is present.

This fixes the issue by replacing viewitems() with items(). The
functionality is the same, the difference is that viewitems() is just a
lightweight view at the contents of the dictionary, whereas items()
creates an actual list of the contents.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>