LoggingCtl: add get_index and prepend option to set_recipe
Method get_index() serves for creating index prefixes for log subfolder
name. Counter for indices is only incremented for the first match of
individual runs.

set_recipe() has been expanded with option prepend, which if is True it
prepends recipe_path with index number in format %02d (00, 01, ...).
With this implementation, every dir name is unique, and is no longer
overriden when one recipe file is specified multiple times in command
line arguments. This also enables alphabetical listing of folders.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>