SlavePool: extend empty pool dir warning message
This patch extends warning message printed when pool directory does not
contain any machines. Now it contains name of the pool and path to the
directory.

This will be used in list_pools implementation, where without this info
the message would not make sense as user wouldn't know which pool is
empty.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>