regression_tests: add 3rd machine to vxlan_multicast
As discussed with Jiri Benc, when using multicast groups with vxlan it's
a good idea to have at least 3 tunnel end points and check functionality
of paralel communication.

This commit adds a guest machine to the first baremetal machine that's
bridged into the same network. This guest makes the 3rd vxlan endpoint.

Another change is the execution of parallel pings from each host to
every other host.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>