lnst-ctl: add -C option to override previously loaded configs
Sometimes we want to ignore automatically loaded config files and just
force use a specific file. I added a new option for this '-C' that
reloads the Controller Config defaults and then loads the specified
config file. It's possible to use the option multiple times (each time
will reset the defaults) and it's also possible to use the '-c' option
after using '-C'.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>