Netperf: add option max_deviation
This option limits the maximum allowed deviation. If the measured
deviation exceeds this limit the result will automatically be FAIL.
This is because sometimes the deviation gets so big that it overlaps
with the specified threshold interval resulting in a PASS even though
the result is so unstable that it shouldn't be compared to anything.

The option accepts either a percentual value describing the maximum
deviation or an absolute value in the same format that we use for
specifying the threshold.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>