regression_tests: add max_deviation to netperf tests
This commit adds a max_deviation parameter to all Netperf test_module
clients in our regression_tests directory. The value of the parameter
can be controlled through alias "nperf_max_dev" that I've added to all
the recipes and set the default value to 20%. I chose the default value
by going through the PASS values we usually see on these tests. None of
these normally exceed 20% so it should be safe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>