NetTest{Slave, Controller}: add is_installed tcpdump check
When the --packet-capture argument was specified and tcpdump wasn't
installed on one of the machines LNST would fail and report a very
nondescriptive exception.

This commit adds a is_installed check to the slave method
"start_packet_capture" and the generated exception is then handled on
the Controller, reporting a more useful error message.

Fixes #161.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>