NetConfigDevice: skip undefined ovs internal port options
When there were no options specified for an ovs internal port, the slave
configuration would fail, with an exception, trying to access a
nonexistant dictionary key. This could be fixe both on the controller
side by sending an empty list of options from the XML parser or on the
slave by checking if the key is defined. However, since we're soon going
to move to pure python recipes I decided to fix it just on the slave
since it's sufficient and the problem won't reappear after the Python
recipes switch.

Fixes #175.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>