Slave: Remove un-nececery tc filter cleanup logic.
There is no need to delete the filter from the ingress qdisc at the
restore flow as the filters will be automatically deleted once the
ingress qdiscs is deleted. Moreover, the syntax of the ingress filters
list has changed in the recent versions of ip which makes this code buggy.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>