NetTestController: add prepare_test_env and init_taskapi methods
prepare_test_env methods purpose is provisioning of the setup and
initializing of TaskAPI objects which represent hosts and interfaces.
It will be used only in upcoming PyRecipes implementation.

init_taskapi just creates instance of TaskAPI class. This instance
is now created in _run_python_task but in upcoming patches it will be
removed from there.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>