recipes: remove ip6 address from bond device in 3 vlans over bond test
The ip6 address was needed to configure 3 vlans over bond setup.
Originally this was a workaround for NM bug. With updated NetworkManager
the ip6 address causes that NM gets stuck on bond device configuration and
never activates it. The solution is to remove the address as the workaround
is no more needed.

Tested with NM: version 1.4

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>