InterfaceManager: Machine: ignore link stats when device not created
When creation and configuration of a device fails, calling the
link_stats method would raise an Exception. Since this is always called
during deconfiguration, a configuration error would result in an
exception related crash of LNST and leave the Slave in an inconsistent
state - unable to run any more recipes unless manually restarted.

This commit fixes the issue by catching the exception and returning an
empty stats dictionary to indicate the failure. This is detected on the
Controller and the logging of device stats is skipped.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>