lnst regression-tests: various bug fixes
Recent removal of the --dump-pools argument in favor of the list_pools
action broke the regression-tests run-tests.sh script.

In addition to that I moved the lnst-ctl function call from the lib.sh
file into the main run-test.sh file and fixed it's handling of the -C vs
-c argument which I previously only fixed in the get_pool_dirs function.

And a final minor change is the addition of the -L argument to rsync
calls so that we also copy linked directories and files such as pools or
slave machine description xmls.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>