InterfaceManager: fix tc deconfiguration
The recently added functionality of clearing tc configuration broke
recipes that already used netem functionality directly from the
configuration XML part. The problem was with trying to clear tc
configuration twice.

To fixed this I moved the new functions to be called after deconfigure()
takes care of it's own tc configuration. For this to work I also had to
add a check for interface existence since deconfigure() can remove the
device completely if NetworkManager is used.

Fixes: 49e523d5 ("Slave: Add basic support for tc.")
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>