recipes: switchdev: Remove IP address from LAG ports
In kernel 4.8, L3 support was added to the mlxsw driver. Therefore,
non-bridged traffic is now directed to the switch itself only if the
packet's DIP matches the interface's IP.

However, due to several pending ordering problems in the switchdev
infrastructure, IP configuration is only offloaded to the device in
cases where it makes sense. Since IP configuration on a LAG port doesn't
make sense the operation is aborted and all the routes in the system are
flushed, thereby causing the machine to become unresponsive.

Until these issues are resolved in the kernel, remove the IP address
from the switch port so that when create_team() is called it won't be
configured upon port being up()-ed.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>