SlavePool: ignore empty parameters on matching
This should help when a user needs to bypass a parameter in requirements.
For example the default value of interface's driver parameter is ixgbe
and the user wants to run on any interface no matter what's in slave pool.
Without this fix it's impossible to do that other than editing the slave
pool xml files directly.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>