HostAPI: Add enable / disable service methods
We are going to need to enable the lldpad service from the QoS recipes,
so we might as well provide generic enable / disable methods for recipes
to use.

In order to support both Systemd and SysVinit systems the requested
service is configured using 'systemctl' or 'service', depending on the
output of `pidof systemd`.

Suggested-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>