InterfaceAPI: Add LLDP support
We can already enable the 'lldpad' daemon from a recipe, but in order to
do anything meaningful with it we need to be able to enable it on an
interface.

Add the if.enable_lldp() method that does just that. During recipe
tear down (graceful or not) make sure that any LLDP-enabled interface is
initialized to default settings. This is done for Ethernet devices only,
as they are the only ones that can be configured using 'lldptool'.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>