NetTestSlave: fix SlaveMethods return values
Due to how the Ctl-Slave communication currently works, returning None
from a Slave Method is not possible - the main Slave loop doesn't resend
this to the Ctl, resulting in the Ctl waiting forever. This is due to
how background and foreground commands work - background command
immediately returns "passed" (meaning started) whereas a forground
command at first returns None, the command then runs in a forked process
and then returns the command results - meanwhile the Ctl is waiting
because it didn't get the "return None" message.

So returning None is fine as long as a different result message is
generated and returned later, currently that only happens in the
run_command method and the others are now fixed by this commit.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>