Slave: deconfigure only created interfaces
Sometimes when configuration of the interface failed during it's
creation (and the device wasn't created, e.g. missing team{d, ctl}). The
slave would register this as a "configured" Device (because it had a
{Net, Nm}ConfigDevice object created, and would attempt to deconfigure
it, this return an exception again and either crash the slave or leave
it in an inconsistent state (trying to deconfigure the device again and
again).

This makes sure that deconfiguration is skipped in case LNST didn't
receive a netlink notification that the device was created and
registered by the kernel.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2: * removed unnecessary conditions in NetTestSlave (suggested by jtluka)
    * instead of using if_index to check device exitence, add a flag the
      describes if the address_setup step was completed without issue -
      could possibly avoid race condition problems of netlink