regression-tests: remove assert_logs from test 28
These checked if stdout from an exec_cmd was printed out in the result
summary. This behaviour changed in 14370a671f614cc728922a13f68963618695c656,
breaking the 28th regression test.

However, these assert_log commands are not needed for the purposes of
this regression test - the required functionality is tested from within
the python task. Because of that it's safe to remove them, in order to
fix the regression test.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>