InterfaceAPI: Configure IP addresses following reset()
Up until commit 82f606fb169e ("{Net, Nm}ConfigDevice: split up() into
up+address_setup") up() used to both set the admin state of an interface
and configure it with IP address(es). However, the mentioned commit
split up() into up() and address_setup() and introduced a regression in
sections of code where up() isn't followed by address_setup().

Fix this by following the call to up() with a call to address_setup() in
reset(), which seems to be the only place where this was missed.

Fixes: 82f606fb169e ("{Net, Nm}ConfigDevice: split up() into up+address_setup")
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>