RecipeCommon: add Offloads module
This module provides functions that are related to Offloads.

The function parse_offloads in the module is intended to be use for parsing
offloads alias. The function parse all used offloads, in each set all offloads
are on, except the ones specified as off.

It takes argument offloads string and returns tuple of used offloads and its
combinations.

It returns list of tuple, containing summary of all used offloads, and list
of all offloads settings.

Expected input: "tso on gso on tx off; tso off tx off ;rx off".

Signed-off-by: Kamil Jerabek <kjerabek@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>