recipes: fix check of interface type for team vs bond recipe
Several phase2 recipes use common python task team_test.py. There's a bug
in the check of interface type that appears when recipe
active_backup_team_vs_active_backup_bond.xml is run.

In this recipe the interface type is bond and the condition matches only
team type and therefore the cpu pinning of NIC irqs is incorrect.

The fix is to add bond type to the condition.

Fixes #180

Signed-off-by: Jan Tluka <jtluka@redhat.com>