Task: HostAPI fix creation of InterfaceAPI objects
The HostAPI class should ignore unused interfaces (specified in slave
config but not matched to the current recipe run), unfortunatelly the
code assumed that these have a None id. This is incorrect as they get a
generated id in the form of "gen_%s_%d".

To fix this we should ignore interface objects that are instances of the
UnusedInterface class.

v2: inverted the condition

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jan Tluka <jtluka@redhat.com>