Task: optimize get_device() function
Function dev_db_get_name() is here called twice and can be reduced to
only one call. Because it queries Slave machine n times, where n is
size of device database on Slave, it can take a lot of time when
database is large.

This reduces time it takes to return from this function
to a half.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>