InterfaceManager: avoid infinite recursion when deconfiguring veth pair
When a user configures veth pair with both veths under root namespace the
infinite recursion is triggered.

Each device of veth pair stores a reference to it's peer in Device's
_master["other"]. It means that LNST will have two veth devices each
referencing the peer as master:

(veth1)                                 (veth2)
Device1._if_index == 10                 Device2._if_index == 20
Device1._master["other"] = [20]         Device2._master["other"] = [10]

On deconfiguration of the veth pair LNST will try to deconfigure
first veth but since it references second veth as master it will start
deconfiguring this master interface. The second interface references the
first interface as master so it will attempt to deconfigure the first veth
device again, and so on until max level of recursion is reached.

This patch adds a check if the device that is deconfigured is referenced by
it's 'other' master. If the reference exists the deconfiguration of master
device will be skipped.

Fixes #187

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>