test_modules/netperf: use bit/s in STREAM tests and parse RR tests
We used kbit/s as output rate for STREAM tests, then we converted it to
bit/s. This patch changes that so the output rate is in bit/s for STREAM
tests and Trans/s for RR tests.

It also adds THROUGHPUT_UNITS omni output selector for omni tests and
adds support for parsing units - both bps and tps

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>