add lnst.Devices
This package contains the implementation of Device
configuration/manipulation classes that are currently supported by LNST.
This implementation has been moved into a separate package that will be
distributed together with the Controller package. These will then be
sent to the Slave during recipe execution so the slave can use them when
required.

The reasoning behind this is that it is often easier to update or make
changes on the Controller than on all the Slaves. Making the Slave
simpler and independent of the implementation of device configuration
means we don't need to update all our slaves as often and also that
changes to the Device configuration implementation don't require any
changes in Slave or Controller code.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2:
    * make Device a Meta class and register RemoteDevice as implementing
      the Device interface
    * removed unnecessary imports
    * added MasterDevice class for slave* methods
    * made master_set to be a 'master' property setter
    * renamed {add, del}_route to route_{add, del}
    * renamed set_speed to speed_set
    * renamed set_autoneg to autoneg_set
    * removed modprobes where unnecessary
    * link_stats are now retrieved from netlink message
    * added link_stats64
    * OvsBridge calls parent _type_init for modprobe
    * remove SoftDevice import - not required