add lnst.Controller.Requirements
This module defines the DeviceReq and HostReq classes, which can be used
to create a global description of Requirements for a network test. You
can use these to define class attributes of a BaseRecipe derived class
to specify "general" requirements for that Recipe, or you can add them
to an instance of a Recipe derived class based on it's parameters to
define requirements "specific" for that single test run.

The module also defines a Requirements class which serves as a container
for HostReq objects, while HostReq classes also serve as containers for
DeviceReq objects. The object tree created this way is translated to a
dictionary used by the internal LNST matching algorithm against
available machines.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>