add lnst.Controller.MachineMapper
Defines the MachineMapper class. Implements a matching algorithm that
maps requirements to available hosts. In this specific class this is
implemented with backtracking, however testers are free to implement
their own algorithm as long as they respect the API of this class as it
needs to integrate with the rest of LNST.

Most of this class is based on the SetupMapper class and includes some
functionality from the old SlavePool class, but has been modified to
work with Python Recipes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2:
    * docstring now explains that implementing your own MachineMapper is
      not supported at this moment
    * removed debug pprint import