add lnst.Controller.Controller
This module defines the Controller class that brings together individual
implementation parts of an LNST Controller. When instantiated, it allows
the tester to configure and run his own recipes with the LNST
'infrastructure'.

This way, the lnst.Controller package is a Python library allowing the
creation of your own LNST Controller.

In theory multiple instances of the Controller class are possible and
you can therefore run multiple recipes in parallel, though this is not
supported and wasn't tested yet.

More info is available in the doc strings of the Controller class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>