add lnst.Tests package
This package will contain all upstream tracked test module classes. At
the moment it only contains the IcmpPing class which is a simple port
from the old IcmpPing test_module. In the future it should be heavily
refactored to support both ipv4 and ipv6 (instead of having 2 different
modules) but for early testing of Python Recipes this is good enough.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>